Issue52-breaking up the account settings form into their own components to practice modular developent
