+++ frontend/src/pages/AccountSettings/AccountSettings.jsx
@@ -1,4 +1,6 @@
-import AccountSettingsForm from '../../components/AccountSettingsForm/AccountSettingsForm';
+import PersonalInfoSection from '../../components/AccountSettings/PersonalInfoSection/PersonalInfoSection';
+import DeleteAccountSection from '../../components/AccountSettings/DeleteAccountSection/DeleteAccountSection';
+import SecuritySection from '../../components/AccountSettings/SecuritySection/SecuritySection';
 import ProfileSidebar from '../../components/ProfileSidebar/ProfileSidebar';
 import styles from './AccountSettings.module.css';
 
@@ -7,7 +9,9 @@ const AccountSettings = () => {
     <>
       <ProfileSidebar />
       <div className={styles["account-settings-div"]}>
-        <AccountSettingsForm />
+        <PersonalInfoSection />
+        <DeleteAccountSection />
+        <SecuritySection /> 
       </div>
     </>
   )
